requirements.txt
atproto==0.0.50
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
schedule==1.2.0
feedparser==6.0.10
Pillow==10.0.1
